chore(codegen): Move code for better coupling (#166)
