+++ codegen/table.go
@@ -4,7 +4,6 @@ import (
 	"fmt"
 	"reflect"
 
-	"github.com/cloudquery/plugin-sdk/schema"
 	"github.com/rs/zerolog"
 	"golang.org/x/exp/slices"
 )
@@ -40,48 +39,6 @@ type (
 	}
 )
 
-func (t *TableDefinition) shouldUnwrapField(field reflect.StructField) bool {
-	switch {
-	case !isFieldStruct(field.Type):
-		return false
-	case slices.Contains(t.structFieldsToUnwrap, field.Name):
-		return true
-	case !field.Anonymous:
-		return false
-	case t.unwrapAllEmbeddedStructFields:
-		return true
-	default:
-		return false
-	}
-}
-
-func (t *TableDefinition) getUnwrappedFields(field reflect.StructField) []reflect.StructField {
-	reflectType := field.Type
-	if reflectType.Kind() == reflect.Ptr {
-		reflectType = reflectType.Elem()
-	}
-
-	fields := make([]reflect.StructField, 0)
-	for i := 0; i < reflectType.NumField(); i++ {
-		sf := reflectType.Field(i)
-		if t.ignoreField(sf) {
-			continue
-		}
-
-		fields = append(fields, sf)
-	}
-	return fields
-}
-
-func isTypeIgnored(t reflect.Type) bool {
-	switch t.Kind() {
-	case reflect.Interface, reflect.Func, reflect.Chan, reflect.UnsafePointer:
-		return true
-	default:
-		return false
-	}
-}
-
 func (t *TableDefinition) ignoreField(field reflect.StructField) bool {
 	switch {
 	case len(field.Name) == 0,
@@ -94,53 +51,8 @@ func (t *TableDefinition) ignoreField(field reflect.StructField) bool {
 	}
 }
 
-func (t *TableDefinition) addColumnFromField(field reflect.StructField, parent *reflect.StructField) error {
-	if t.ignoreField(field) {
-		return nil
-	}
-
-	columnType, err := t.typeTransformer(field)
-	if err != nil {
-		return fmt.Errorf("failed to transform type for field %s: %w", field.Name, err)
-	}
-
-	if columnType == schema.TypeInvalid {
-		columnType, err = DefaultTypeTransformer(field)
-		if err != nil {
-			return fmt.Errorf("failed to transform type for field %s: %w", field.Name, err)
-		}
-	}
-
-	// generate a PathResolver to use by default
-	pathResolver := fmt.Sprintf(`schema.PathResolver("%s")`, field.Name)
-	name, err := t.nameTransformer(field)
-	if err != nil {
-		return fmt.Errorf("failed to transform field name for field %s: %w", field.Name, err)
-	}
-	// skip field if there is no name
-	if name == "" {
-		return nil
-	}
-	if parent != nil {
-		pathResolver = fmt.Sprintf(`schema.PathResolver("%s.%s")`, parent.Name, field.Name)
-		parentName, err := t.nameTransformer(*parent)
-		if err != nil {
-			return fmt.Errorf("failed to transform field name for parent field %s: %w", parent.Name, err)
-		}
-		name = fmt.Sprintf("%s_%s", parentName, name)
-	}
-
-	column := ColumnDefinition{
-		Name:     name,
-		Type:     columnType,
-		Resolver: pathResolver,
-	}
-	t.Columns = append(t.Columns, column)
-	return nil
-}
-
 // NewTableFromStruct creates a new TableDefinition from a struct by inspecting its fields
-func NewTableFromStruct(name string, obj interface{}, opts ...TableOptions) (*TableDefinition, error) {
+func NewTableFromStruct(name string, obj interface{}, opts ...TableOption) (*TableDefinition, error) {
 	t := &TableDefinition{
 		Name:            name,
 		nameTransformer: DefaultNameTransformer,
@@ -160,22 +72,16 @@ func NewTableFromStruct(name string, obj interface{}, opts ...TableOptions) (*Ta
 
 	t.Columns = append(t.Columns, t.extraColumns...)
 
+	eType := e.Type()
 	for i := 0; i < e.NumField(); i++ {
-		field := e.Type().Field(i)
+		field := eType.Field(i)
 
-		if t.shouldUnwrapField(field) {
-			unwrappedFields := t.getUnwrappedFields(field)
-			var parent *reflect.StructField
-			// For non embedded structs we need to add the parent field name to the path
-			if !field.Anonymous {
-				parent = &field
+		switch {
+		case t.shouldUnwrapField(field):
+			if err := t.unwrapField(field); err != nil {
+				return nil, err
 			}
-			for _, f := range unwrappedFields {
-				if err := t.addColumnFromField(f, parent); err != nil {
-					return nil, fmt.Errorf("failed to add column from field %s: %w", f.Name, err)
-				}
-			}
-		} else {
+		default:
 			if err := t.addColumnFromField(field, nil); err != nil {
 				return nil, fmt.Errorf("failed to add column for field %s: %w", field.Name, err)
 			}
