+++ codegen/table_test.go
@@ -12,131 +12,127 @@ import (
 	"github.com/google/go-cmp/cmp/cmpopts"
 )
 
-type embeddedStruct struct {
-	EmbeddedString string
-}
-
-type testStruct struct {
-	// IntCol this is an example documentation comment
-	IntCol    int     `json:"int_col,omitempty"`
-	StringCol string  `json:"string_col,omitempty"`
-	FloatCol  float64 `json:"float_col,omitempty"`
-	BoolCol   bool    `json:"bool_col,omitempty"`
-	JSONCol   struct {
-		IntCol    int    `json:"int_col,omitempty"`
-		StringCol string `json:"string_col,omitempty"`
+type (
+	embeddedStruct struct {
+		EmbeddedString string
 	}
-	IntArrayCol    []int      `json:"int_array_col,omitempty"`
-	StringArrayCol []string   `json:"string_array_col,omitempty"`
-	TimeCol        time.Time  `json:"time_col,omitempty"`
-	TimePointerCol *time.Time `json:"time_pointer_col,omitempty"`
-	JSONTAG        *string    `json:"json_tag"`
-	SKIPJSONTAG    *string    `json:"-"`
-	NOJSONTAG      *string
-	InterfaceCol   interface{}
 
-	*embeddedStruct
-}
+	testStruct struct {
+		// IntCol this is an example documentation comment
+		IntCol    int     `json:"int_col,omitempty"`
+		StringCol string  `json:"string_col,omitempty"`
+		FloatCol  float64 `json:"float_col,omitempty"`
+		BoolCol   bool    `json:"bool_col,omitempty"`
+		JSONCol   struct {
+			IntCol    int    `json:"int_col,omitempty"`
+			StringCol string `json:"string_col,omitempty"`
+		}
+		IntArrayCol    []int      `json:"int_array_col,omitempty"`
+		StringArrayCol []string   `json:"string_array_col,omitempty"`
+		TimeCol        time.Time  `json:"time_col,omitempty"`
+		TimePointerCol *time.Time `json:"time_pointer_col,omitempty"`
+		JSONTAG        *string    `json:"json_tag"`
+		SKIPJSONTAG    *string    `json:"-"`
+		NOJSONTAG      *string
+		*embeddedStruct
+	}
+	testStructWithEmbeddedStruct struct {
+		*testStruct
+		*embeddedStruct
+	}
+	testStructWithNonEmbeddedStruct struct {
+		TestStruct  *testStruct
+		NonEmbedded *embeddedStruct
+	}
+	testStructWithCustomType struct {
+		TimeCol time.Time `json:"time_col,omitempty"`
+	}
+)
 
-type testStructWithEmbeddedStruct struct {
-	*testStruct
-	*embeddedStruct
-}
-
-type testStructWithNonEmbeddedStruct struct {
-	TestStruct  *testStruct
-	NonEmbedded *embeddedStruct
-}
-
-type testStructWithCustomType struct {
-	TimeCol time.Time `json:"time_col,omitempty"`
-}
-
-var expectedColumns = []ColumnDefinition{
-	{
-		Name:     "int_col",
-		Type:     schema.TypeInt,
-		Resolver: `schema.PathResolver("IntCol")`,
-	},
-	{
-		Name:     "string_col",
-		Type:     schema.TypeString,
-		Resolver: `schema.PathResolver("StringCol")`,
-	},
-	{
-		Name:     "float_col",
-		Type:     schema.TypeFloat,
-		Resolver: `schema.PathResolver("FloatCol")`,
-	},
-	{
-		Name:     "bool_col",
-		Type:     schema.TypeBool,
-		Resolver: `schema.PathResolver("BoolCol")`,
-	},
-	{
-		Name:     "json_col",
-		Type:     schema.TypeJSON,
-		Resolver: `schema.PathResolver("JSONCol")`,
-	},
-	{
-		Name:     "int_array_col",
-		Type:     schema.TypeIntArray,
-		Resolver: `schema.PathResolver("IntArrayCol")`,
-	},
-	{
-		Name:     "string_array_col",
-		Type:     schema.TypeStringArray,
-		Resolver: `schema.PathResolver("StringArrayCol")`,
-	},
-	{
-		Name:     "time_col",
-		Type:     schema.TypeTimestamp,
-		Resolver: `schema.PathResolver("TimeCol")`,
-	},
-	{
-		Name:     "time_pointer_col",
-		Type:     schema.TypeTimestamp,
-		Resolver: `schema.PathResolver("TimePointerCol")`,
-	},
-	{
-		Name:     "json_tag",
-		Type:     schema.TypeString,
-		Resolver: `schema.PathResolver("JSONTAG")`,
-	},
-	{
-		Name:     "nojsontag",
-		Type:     schema.TypeString,
-		Resolver: `schema.PathResolver("NOJSONTAG")`,
-	},
-}
-
-var expectedTestTable = TableDefinition{
-	Name:            "test_struct",
-	Columns:         expectedColumns,
-	nameTransformer: DefaultNameTransformer,
-}
-
-var expectedTestTableEmbeddedStruct = TableDefinition{
-	Name:            "test_struct",
-	Columns:         append(expectedColumns, ColumnDefinition{Name: "embedded_string", Type: schema.TypeString, Resolver: `schema.PathResolver("EmbeddedString")`}),
-	nameTransformer: DefaultNameTransformer,
-}
-
-var expectedTestTableNonEmbeddedStruct = TableDefinition{
-	Name: "test_struct",
-	Columns: ColumnDefinitions{
-		// Should not be unwrapped
-		ColumnDefinition{Name: "test_struct", Type: schema.TypeJSON, Resolver: `schema.PathResolver("TestStruct")`},
-		// Should be unwrapped
-		ColumnDefinition{Name: "non_embedded_embedded_string", Type: schema.TypeString, Resolver: `schema.PathResolver("NonEmbedded.EmbeddedString")`},
-	},
-	nameTransformer: DefaultNameTransformer,
-}
+var (
+	expectedColumns = []ColumnDefinition{
+		{
+			Name:     "int_col",
+			Type:     schema.TypeInt,
+			Resolver: `schema.PathResolver("IntCol")`,
+		},
+		{
+			Name:     "string_col",
+			Type:     schema.TypeString,
+			Resolver: `schema.PathResolver("StringCol")`,
+		},
+		{
+			Name:     "float_col",
+			Type:     schema.TypeFloat,
+			Resolver: `schema.PathResolver("FloatCol")`,
+		},
+		{
+			Name:     "bool_col",
+			Type:     schema.TypeBool,
+			Resolver: `schema.PathResolver("BoolCol")`,
+		},
+		{
+			Name:     "json_col",
+			Type:     schema.TypeJSON,
+			Resolver: `schema.PathResolver("JSONCol")`,
+		},
+		{
+			Name:     "int_array_col",
+			Type:     schema.TypeIntArray,
+			Resolver: `schema.PathResolver("IntArrayCol")`,
+		},
+		{
+			Name:     "string_array_col",
+			Type:     schema.TypeStringArray,
+			Resolver: `schema.PathResolver("StringArrayCol")`,
+		},
+		{
+			Name:     "time_col",
+			Type:     schema.TypeTimestamp,
+			Resolver: `schema.PathResolver("TimeCol")`,
+		},
+		{
+			Name:     "time_pointer_col",
+			Type:     schema.TypeTimestamp,
+			Resolver: `schema.PathResolver("TimePointerCol")`,
+		},
+		{
+			Name:     "json_tag",
+			Type:     schema.TypeString,
+			Resolver: `schema.PathResolver("JSONTAG")`,
+		},
+		{
+			Name:     "nojsontag",
+			Type:     schema.TypeString,
+			Resolver: `schema.PathResolver("NOJSONTAG")`,
+		},
+	}
+	expectedTestTable = TableDefinition{
+		Name:            "test_struct",
+		Columns:         expectedColumns,
+		nameTransformer: DefaultNameTransformer,
+	}
+	expectedTestTableEmbeddedStruct = TableDefinition{
+		Name:            "test_struct",
+		Columns:         append(expectedColumns, ColumnDefinition{Name: "embedded_string", Type: schema.TypeString, Resolver: `schema.PathResolver("EmbeddedString")`}),
+		nameTransformer: DefaultNameTransformer,
+	}
+	expectedTestTableNonEmbeddedStruct = TableDefinition{
+		Name: "test_struct",
+		Columns: ColumnDefinitions{
+			// Should not be unwrapped
+			ColumnDefinition{Name: "test_struct", Type: schema.TypeJSON, Resolver: `schema.PathResolver("TestStruct")`},
+			// Should be unwrapped
+			ColumnDefinition{Name: "non_embedded_embedded_string", Type: schema.TypeString, Resolver: `schema.PathResolver("NonEmbedded.EmbeddedString")`},
+		},
+		nameTransformer: DefaultNameTransformer,
+	}
+)
 
 func TestTableFromGoStruct(t *testing.T) {
 	type args struct {
 		testStruct interface{}
-		options    []TableOptions
+		options    []TableOption
 	}
 
 	tests := []struct {
@@ -155,7 +151,7 @@ func TestTableFromGoStruct(t *testing.T) {
 			name: "should unwrap all embedded structs when option is set",
 			args: args{
 				testStruct: testStructWithEmbeddedStruct{},
-				options:    []TableOptions{WithUnwrapAllEmbeddedStructs()},
+				options:    []TableOption{WithUnwrapAllEmbeddedStructs()},
 			},
 			want: expectedTestTableEmbeddedStruct,
 		},
@@ -163,7 +159,7 @@ func TestTableFromGoStruct(t *testing.T) {
 			name: "should unwrap specific structs when option is set",
 			args: args{
 				testStruct: testStructWithNonEmbeddedStruct{},
-				options:    []TableOptions{WithUnwrapFieldsStructs([]string{"NonEmbedded"})},
+				options:    []TableOption{WithUnwrapStructFields([]string{"NonEmbedded"})},
 			},
 			want: expectedTestTableNonEmbeddedStruct,
 		},
@@ -171,7 +167,7 @@ func TestTableFromGoStruct(t *testing.T) {
 			name: "should override schema type when option is set",
 			args: args{
 				testStruct: testStructWithCustomType{},
-				options: []TableOptions{WithTypeTransformer(func(t reflect.StructField) (schema.ValueType, error) {
+				options: []TableOption{WithTypeTransformer(func(t reflect.StructField) (schema.ValueType, error) {
 					switch t.Type {
 					case reflect.TypeOf(time.Time{}), reflect.TypeOf(&time.Time{}):
 						return schema.TypeJSON, nil
